FEATURE: Implemented STATUS command and de-debugged timestamps
The new STATUS command currently prints the Rover's remaining energy.
In the future it will provide more information.  The timestamps that
used to be clumsily printed out as DEBUG messages are now elegantly
incorporated into the rover command prompt.  Several sections of
defunct code in GameEngine.java had to be commented out in order to
keep GridRover buildable, due to the other changes that were made.
That code will be removed in an upcoming cleanup commit anyway.