FRAMEWORK: Implemented xml file reading and began reorganizing initialization
This was a big commit!  I started reorganizing the initialization to make
more sense.  When I'm done, I think the GameEngine will be sleeker and
more flexible.  I also wrote a small xml file with a single datum for a
physical object inside.  I added code to GridRover.java to read this file
and use the data inside to make an ArrayList of physical objects which
can later be used as prototypes for items to be littered across the map
for the rover to pick up.  Added some TODOs.