FEATURE: Items read from file are now scattered across map
In this commit, I rounded off the changes to initialization that started
last commit.  Added a method to the MapGrid for scattering random items
across the map.  It makes use of a new Item constructor and is made
available by the new GameEngine method scatterItemsRandomly.  Together,
these changes mean that GridRover now scatters items read from file
across the game map, making this a milestone commit.